requirements.txt
Flask
Flask-Cors
pymongo
requests
python-dotenv
